requirements.txt
langchain
ctransformers
chainlit
langchain_community
openai-whisper
transformers